Module_1_Python/Lesson_10_ORM_SQLAlchemy_Project_part_1_of_4/requirements.txt
fastapi==0.103.1
sqlalchemy==1.4.39
pydantic==2.3.0
python_dotenv==1.0.0